requirements.txt
feedparser>=6.0.0
PyYAML>=6.0
python-dateutil>=2.8.0
